+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/dsl/query/SqlQueryEntityProperties.java
@@ -230,5 +230,4 @@ public class SqlQueryEntityProperties extends AbstractSqlQueryEntityProperties<S
     public QuerySortExpression sort() {
         return new SqlQueryExpression(jdbc, classMapping, selectBuilder).sort();
     }
-
 }